Ensure methyldackel process runs on each sample.
Before this patch, the methyldackel process would receive inputs from the channels:
- ch_bam_dedup_for_methyldackel (containing many bams)
- ch_bam_index_for_methyldackel (containing many bais)
- ch_fasta_for_methyldackel (containing only one fasta)
- ch_fasta_index_for_methyldackel (containing only one fai)

The process would only run once as the ch_fasta_for_methyldackel and ch_fasta_index_for_methyldackel channels would be quickly exhausted.

This patch uses the combine operator to ensure that every bam+bai pair runs through the methyldackel process. I've noticed that this style of nextflow process coordination is not used very often in nf-core, so there is a chance the PR violates nf-core style guidelines. There were no indications of that from nf-core lint, but appreciate that the PR style might not be appropriate.