Update methylseq.nf
Include module twice: Qualimap (Before and After Markduplicates in the BAM file) Add module Qualimap into the workflow twice (QUALIMAP and QUALIMAP2 

The syntax of markduplicates module was corrected. 

HS metrics was also corrected with an else statment to avoid the use of the module in case that BED file is not included as an input.